+++ oap-server/server-query-plugin/query-graphql-plugin/src/main/java/org/apache/skywalking/oap/query/graphql/GraphQLQueryHandler.java
@@ -24,19 +24,37 @@ import com.linecorp.armeria.server.ServiceRequestContext;
 import com.linecorp.armeria.server.annotation.Blocking;
 import com.linecorp.armeria.server.annotation.Post;
 import com.linecorp.armeria.server.graphql.GraphqlService;
+import graphql.analysis.MaxQueryComplexityInstrumentation;
 import graphql.schema.GraphQLSchema;
+import lombok.extern.slf4j.Slf4j;
 
+@Slf4j
 public class GraphQLQueryHandler {
     private final GraphqlService graphqlService;
 
-    public GraphQLQueryHandler(final GraphQLSchema schema) {
-        graphqlService = GraphqlService.builder().schema(schema).build();
+    public GraphQLQueryHandler(
+        final GraphQLQueryConfig config,
+        final GraphQLSchema schema) {
+        final int allowedComplexity = config.getMaxQueryComplexity();
+        graphqlService =
+            GraphqlService
+                .builder()
+                .schema(schema)
+                .instrumentation(new MaxQueryComplexityInstrumentation(allowedComplexity, info -> {
+                    log.warn(
+                        "Aborting query because it's too complex, maximum allowed is [{}] but was [{}]",
+                        allowedComplexity,
+                        info.getComplexity());
+                    return true;
+                }))
+                .build();
     }
 
     @Blocking
     @Post("/graphql")
-    public HttpResponse graphql(ServiceRequestContext ctx,
-                                HttpRequest req) throws Exception {
+    public HttpResponse graphql(
+        final ServiceRequestContext ctx,
+        final HttpRequest req) throws Exception {
         return graphqlService.serve(ctx, req);
     }
 }