+++ oap-server/server-query-plugin/query-graphql-plugin/src/main/java/org/apache/skywalking/oap/query/graphql/GraphQLQueryProvider.java
@@ -126,7 +126,8 @@ public class GraphQLQueryProvider extends ModuleProvider {
         HTTPHandlerRegister service = getManager().find(CoreModule.NAME)
                                                   .provider()
                                                   .getService(HTTPHandlerRegister.class);
-        service.addHandler(new GraphQLQueryHandler(schemaBuilder.build().makeExecutableSchema()));
+        service.addHandler(
+            new GraphQLQueryHandler(config, schemaBuilder.build().makeExecutableSchema()));
     }
 
     @Override